% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queueing_network.R
\name{queueing_network_poi_geom}
\alias{queueing_network_poi_geom}
\title{Simulate a Two-Node Queueing Network}
\usage{
queueing_network_poi_geom(n_obs, burn_in, p_12, p_21, lambda_1, lambda_2, G_1,
  G_2)
}
\arguments{
\item{n_obs}{Integer, number of observations to be analyzed}

\item{burn_in}{Integer, length of burn in phase in order to get a stationary 
distribution. Netowrk starts empty}

\item{p_12}{Numeric, in [0,1], routing probability from node 1 to node 2}

\item{p_21}{Numeric, in [0,1], routing probability from node 2 to node 1}

\item{lambda_1}{Numeric (> 0), (Poisson) arrival rate at node 1}

\item{lambda_2}{Numeric (> 0), (Poisson) arrival rate at node 2}

\item{G_1}{Numeric (in [0,1]), parameter steering the (Geometric) service 
time distribution at node 1}

\item{G_2}{Numeric (in [0,1]), parameter steering the (Geometric) service 
time distribution at node 2}
}
\value{
Data Frame with 4 columns (number of arrivals and departures) with 
n_obs + burn_in rows, each representing one time slot.
}
\description{
Creates a simulation of a Poi/Geom/infty-Network with 2 nodes and routing 
probabilities p_12 and p_21.
}
